fix: 修复实体与数据库字段匹配问题，修正服务层类型转换
主要修复：
- TestCase: 完整重写，添加数据库缺失字段，添加兼容字段解决BaseEntity冲突
- TestExecution: 添加 browser、appVersion 字段
- ScheduledTask: 添加 caseIds、totalExecutions 等统计字段
- Project: 添加 creatorId、updaterId、version 数据库字段映射
- ApiRequest: 重命名 version 为 apiVersion 避免与 BaseEntity 冲突
- ApiCollection: 添加 getDeleted() 兼容方法

服务层修复：
- 修复 Long/Integer 类型转换问题
- 修复 selectByTestSuiteId 参数类型
- ApiClient.ApiResponse 添加 url 字段

+++ iatms-v2/src/main/java/com/iatms/application/testing/TestSuiteCommandService.java
@@ -52,7 +52,7 @@ public class TestSuiteCommandService {
         testSuiteMapper.updateById(testSuite);
 
         // 更新关联的用例
-        testSuiteRequestMapper.deleteByTestSuiteId(id);
+        testSuiteRequestMapper.deleteByTestSuiteId(id.intValue());
         if (caseIds != null && !caseIds.isEmpty()) {
             for (Long caseId : caseIds) {
                 TestSuiteRequest link = new TestSuiteRequest();
@@ -69,7 +69,7 @@ public class TestSuiteCommandService {
     @Transactional
     public void delete(Long id) {
         testSuiteMapper.deleteById(id);
-        testSuiteRequestMapper.deleteByTestSuiteId(id);
+        testSuiteRequestMapper.deleteByTestSuiteId(id.intValue());
         log.info("删除测试套件: {}", id);
     }
